Updated to libcouchbase 3.1.2.
Reviewed-on: http://review.couchbase.org/c/couchnode/+/152508
Tested-by: Brett Lawson <brett19@gmail.com>

+++ deps/lcb/contrib/cJSON/cJSON.h
@@ -23,6 +23,8 @@
 #ifndef cJSON__h
 #define cJSON__h
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C"
 {
@@ -72,7 +74,7 @@ typedef struct cJSON {
         int type; /* The type of the item, as above. */
 
         char *valuestring; /* The item's string, if type==cJSON_String */
-        int valueint; /* The item's number, if type==cJSON_Number */
+        int64_t valueint;  /* The item's number, if type==cJSON_Number */
         double valuedouble; /* The item's number, if type==cJSON_Number */
 
         char *string; /* The item's name string, if this item is the
